AgentOps/single-agent-monitoring/pyproject.toml
[project]
name = "single-agent-monitoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agentops>=0.4.18",
    "openai-agents>=0.2.4",
    "python-dotenv>=1.1.1",
]
